+++ resources/lang/en/authority.php
@@ -55,4 +55,5 @@ return array(
     'authors_found'       => 'artists found for',
     'authors_counted'     => 'artists',
     'authors_none'        => 'currently no artists',
+    'roles'               => 'roles',
 );